reject empty user ids on delete

+++ src/main/java/com/sinnerschrader/skillwill/services/UserService.java
@@ -140,7 +140,9 @@ public class UserService {
   }
 
   public void deleteUserById(String id) {
-
+    if (StringUtils.isEmpty(id)) {
+      throw new EmptyArgumentException("");
+    }
     userRepository.deleteById(id);
   }
 }
